Added some gameplay logic
Basic stuff right now - just some if statements and loops to check things, but most of the logic behind it isn't there yet. Just output etc. Might also need to change some of the logic behind deck (as in - have a method to actually take a card from the top) unless I change it back to just a CardCollection attribute.