Adding round logic
All subject to change. Logic here is that when a player inputs to end their turn, there is a method that handles moving all cards in the play area to bank (and then clearing the play area). Now, I think I need to focus on card logic and adding all child classes