Clean up
What I had to do here was move everything to a new project. I think I had created it wrong or initialised the repository wrong and suddenly many errors popped up that I could not solve. Now the problem seems to be fixed. All the code was the same as before