Modified end turn and child card classes
End turn now marks each card with will add to bank (currently no implementation). The idea here is that it should only do anything to chest and key (so check if there is chest/key also being banked). It can be used on the other cards, but won't do anything so have added an empty base implementation that can be overridden by any classes needing it