Implemented chest and key logic
Needed to add draw card method to discard pile. I decided to just have the checking logic in chest - because there isn't much need to have it in both - and it might cause the count/draw to happen twice, which is not what I want.  Basically, we get the cards from the play area, count them, and loop based on that size, drawing a card each iteration. The output is a little funny, and may need to be fixed. Have not got to the stage of testing the program function - but getting close